requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
django-filter
djangorestframework
python-dateutil
pytz
requests
simplejson
six
sqlparse
urllib3
gunicorn
django-hashid-field
psycopg2
django-crontab
whitenoise
django-heroku